pyproject.toml
[project]
name = "cleanstack"
version = "0.14.0"
requires-python = ">=3.14"
dependencies = [
    "faker>=40.8",
    "fastapi>=0.135",
    "pydantic-settings>=2.13",
]

[project.optional-dependencies]
mongo = [
    "pymongo>=4.16",
]
sql = [
    "psycopg[binary]>=3.3",
    "sqlalchemy>=2.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28",
    "mypy>=1.19",
    "pre-commit>=4.5",
    "pytest>=9.0",
    "ruff>=0.15",
    "uvicorn>=0.41.0",
]

[tool.setuptools.packages.find]
include = ["cleanstack*"]

[tool.ruff]
output-format = "concise"
[tool.ruff.lint]
select = [
    "E", "W", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
    "PL", # Pylint
    "RUF",
]
ignore = [
    "PLR0911", # Too many return statements (7 > 6)
    "PLR0913", # Too many arguments in function definition (6 > 5)
]

[tool.ty.terminal]
output-format = "concise"

[tool.mypy]
strict = true
plugins = ['pydantic.mypy']

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "info"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)30s - %(message)s"
